# qa-intermediate/backend-project/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Finanzas Personales API
  description: |
    API REST para gestionar transacciones financieras y categorización de gastos e ingresos.
    
    ## Funcionalidades principales:
    - **Gestión de Categorías**: Crear y consultar categorías de transacciones
    - **Gestión de Transacciones**: Registrar transacciones con categorización opcional
    - **Consultas Avanzadas**: Filtros por categoría, tipo, fechas, montos y búsqueda de texto
    - **Agregaciones**: Resúmenes agrupados por categoría con cálculos netos
    - **Paginación**: Soporte completo de paginación y ordenamiento
    
    ## Convenciones:
    - **CREDIT**: Ingresos (suma positiva)
    - **DEBIT**: Egresos (suma negativa)
    - **Zona horaria**: America/Lima para agregaciones por fecha
    - **Formato de fechas**: ISO-8601 (yyyy-MM-dd) para filtros de fecha
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: desarrollo@interbank.pe
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api-finanzas.interbank.pe
    description: Servidor de producción
paths:
  /api/categories:
    post:
      tags:
        - Categorías
      summary: Crear nueva categoría
      description: Crea una nueva categoría de transacciones con validación de nombre único
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
            examples:
              alimentacion:
                summary: Categoría de Alimentación
                value:
                  name: "Alimentación"
                  colorHex: "#FF5722"
              transporte:
                summary: Categoría de Transporte
                value:
                  name: "Transporte"
                  colorHex: "#2196F3"
      responses:
        '201':
          description: Categoría creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
              examples:
                created:
                  summary: Categoría creada
                  value:
                    id: 1
                    name: "Alimentación"
                    colorHex: "#FF5722"
                    createdAt: "2025-09-15T10:30:00Z"
        '400':
          description: Error de validación o categoría duplicada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  summary: Error de validación
                  value:
                    timestamp: "2025-09-15T10:30:00Z"
                    status: 400
                    error: "Validation Failed"
                    message: "Error de validación en los datos de entrada"
                    details:
                      name: "El nombre de la categoría es obligatorio"
                duplicate_name:
                  summary: Nombre duplicado
                  value:
                    timestamp: "2025-09-15T10:30:00Z"
                    status: 400
                    error: "Bad Request"
                    message: "Ya existe una categoría con el nombre: Alimentación"

    get:
      tags:
        - Categorías
      summary: Listar categorías
      description: |
        Obtiene una lista paginada de categorías con filtros opcionales.
        
        **Parámetros de ordenamiento permitidos:**
        - `name` (por defecto)
        - `createdAt`
        
        **Formato:** `campo,dirección` (ej: `name,ASC` o `createdAt,DESC`)
      operationId: getCategories
      parameters:
        - name: q
          in: query
          description: Búsqueda por nombre de categoría (parcial, case-insensitive)
          required: false
          schema:
            type: string
          example: "aliment"
        - name: page
          in: query
          description: Número de página (basado en 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
        - name: size
          in: query
          description: Cantidad de elementos por página (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: sort
          in: query
          description: Campo y dirección de ordenamiento
          required: false
          schema:
            type: string
            pattern: '^(name|createdAt),(ASC|DESC)$'
            default: "name,ASC"
          example: "name,ASC"
      responses:
        '200':
          description: Lista de categorías obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCategoryDto'
              examples:
                categories_page:
                  summary: Página de categorías
                  value:
                    content:
                      - id: 1
                        name: "Alimentación"
                        colorHex: "#FF5722"
                        createdAt: "2025-09-15T10:30:00Z"
                      - id: 2
                        name: "Transporte"
                        colorHex: "#2196F3"
                        createdAt: "2025-09-15T10:35:00Z"
                    pageable:
                      pageNumber: 0
                      pageSize: 20
                      sort:
                        sorted: true
                        ascending: true
                    totalElements: 2
                    totalPages: 1
                    first: true
                    last: true
                    numberOfElements: 2
        '400':
          description: Parámetros de consulta inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/categories/{id}:
    get:
      tags:
        - Categorías
      summary: Obtener categoría por ID
      description: Obtiene los detalles de una categoría específica por su identificador
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          description: ID único de la categoría
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Categoría encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
              examples:
                category:
                  summary: Detalles de categoría
                  value:
                    id: 1
                    name: "Alimentación"
                    colorHex: "#FF5722"
                    createdAt: "2025-09-15T10:30:00Z"
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: Categoría no encontrada
                  value:
                    timestamp: "2025-09-15T10:30:00Z"
                    status: 400
                    error: "Bad Request"
                    message: "Categoría no encontrada con ID: 999"
...
components:
  schemas:
    CreateCategoryDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: Nombre único de la categoría
          example: "Alimentación"
        colorHex:
          type: string
          maxLength: 7
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Color hexadecimal para la categoría (formato #RRGGBB)
          example: "#FF5722"
      description: Datos requeridos para crear una nueva categoría
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único de la categoría
          example: 1
        name:
          type: string
          description: Nombre de la categoría
          example: "Alimentación"
        colorHex:
          type: string
          description: Color hexadecimal de la categoría
          example: "#FF5722"
        createdAt:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2025-09-15T10:30:00Z"
      description: Información completa de una categoría
...
    PageCategoryDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
          description: Lista de categorías en la página actual
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          description: Total de elementos en todas las páginas
          example: 50
        totalPages:
          type: integer
          description: Total de páginas disponibles
          example: 3
        first:
          type: boolean
          description: Indica si es la primera página
          example: true
        last:
          type: boolean
          description: Indica si es la última página
          example: false
        numberOfElements:
          type: integer
          description: Número de elementos en la página actual
          example: 20
      description: Respuesta paginada de categorías
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
          description: Número de página actual (basado en 0)
          example: 0
        pageSize:
          type: integer
          description: Tamaño de la página
          example: 20
        sort:
          $ref: '#/components/schemas/Sort'
      description: Información de paginación
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
          description: Indica si los resultados están ordenados
          example: true
        ascending:
          type: boolean
          description: Indica si el orden es ascendente
          example: false
      description: Información de ordenamiento
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento en que ocurrió el error
          example: "2025-09-15T10:30:00Z"
        status:
          type: integer
          description: Código de estado HTTP
          example: 400
        error:
          type: string
          description: Tipo de error
          example: "Bad Request"
        message:
          type: string
          description: Mensaje descriptivo del error
          example: "Error de validación en los datos de entrada"
        details:
          type: object
          additionalProperties:
            type: string
          description: Detalles específicos del error (ej. campos de validación)
          example:
            name: "El nombre de la categoría es obligatorio"
            amount: "El monto debe ser mayor a 0"
      description: Estructura estándar para respuestas de error
...